[ttFont] Split TTGlyphSet code into ttGlyphSet.py
